model backup

+++ util.py
@@ -23,7 +23,7 @@ def get_graph(grp_id:int,window:int) :
 
 def get_output(grp_id:int,window:int):
     df = pd.read_csv("target_class_windowwise/"+str(grp_id) + "_" + str(window) + ".csv")
-    return torch.tensor(df.iloc[0,:]).float()
+    return torch.tensor(df.iloc[0,:])
 
 def get_member_role_vectors(grp_id:int,window:int):
     df = pd.read_csv("NMF_member_level_features_windowwise/"+str(grp_id) + "_" + str(window) + ".csv")
